Form added and app.js updated
In app.js express, https, and body parser were installed and required. It now has a .get, .post to the app and an API call to openweathermaps.org via https.get. An input is now being taken from the index.html via the new form added in and used to call the corresponding city data. The presentation is still rudimentary, though it will be updated.